EES-4497 Rename TestContact mock data
